New item structure basically finished
Items are now a collection of item components, which are moddable and defined in mod scripts. Next steps should be to allow item components to bear some kind of actual functionality other than data, so like references to an IUseTool-implementation which will be used when using a tool. Also start work on interactable entities, an interactable component which specifies HOW a player interacts with the entity, as well as what happens when it is interacted with.