Client entity creation prediction working very crudely
Currently, there are cases where there appears to occur some kind of data race - the resulting odd behaviour is that two identical entities are created. I am assuming that one of them is the predicted one, and the other being the server side one. Needs to be further investigated to make sure that it doesn't happen when placing entities, which will be central to the game's mechanics.