Added some scripting API for sending container provider data, containers can now send custom data from the server to the client
This functionality should primarily be used for visual cues etc. like loading bars in machines. DO NOT send reliable data from the server to a client using it, instead it should be snapshots of data that otherwise can be predicted on the client. Also, try to make sure that you do not send these snapshots too often, will cause congestion.