Refactored entire ECS protocol and refactored scripting
Whenever a property inside a component updates, only that property is packed up and sent over the network, this greatly reduces bandwidth and is more in line with the real way you should delta compress. Also, this means that every property now must have its own Packer which defines how the type of that property gets packed into a byte array and unpacked from a byte array. Interpolation is also done per-property now. All of this is specified in the ComponentPropertyAttribute which is required for all component properties. Scripting has also been refactored a bit. Types defined in mods must now had the  attribute to propagate to the game.